update from Ryan:
I've just completed a bunch of updates on the PHP API. A few bugs were
fixed, most notably a problem with large binary files being stored. The
incr and decr commands are supported for numeric values. Better error
handling has been implemented and more descriptive details are given
when debug is on.

I've made a bunch of updates as I noted in the email sent to the list.
Also I've moved the Documentation out of the source and into its own
file. And I've added a changelog to keep track of the updates.

Ryan


git-svn-id: http://code.sixapart.com/svn/memcached@74 b0b603af-a30f-0410-a34e-baf09ae79d0b