Brad
attached is the patch to move the 1.0.7 php api to 1.0.8. it fixes the
incorrect debug wording Brion mentioned earlier and also incorporates
his patch that for the debugging abstraction. It also fixes some of the
formatting problems, my editor seemed to use spaces where tabs should
have been applied in some areas. I might consider editing on emacs or vi
in the future.

I'll work on the _load_items() function for the next release and see
what I can do about it.

Ryan


git-svn-id: http://code.sixapart.com/svn/memcached@92 b0b603af-a30f-0410-a34e-baf09ae79d0b