Refactor the command parsing. We now break incoming command lines up into tokens and operate on lists of tokens rather than on raw strings. Since we know how long all the tokens are by virtue of scanning the command line for them, this allows a lot of strlen() calls to be eliminated, principally in assoc.c, in favor of passing key lengths in as a parameter.
In the course of making this change, the command processing in memcached.c
was broken into separate functions for each command. That makes it much
easier to get meaningful results from function-level profiling.

This includes various other minor tweaks such as changing the order of the
tests for command keywords so that "get" comes first (since it's the most
frequent command).


git-svn-id: http://code.sixapart.com/svn/memcached@419 b0b603af-a30f-0410-a34e-baf09ae79d0b